Update codes to apply business logic
