+++ server/db/serializers/market.py
@@ -2,7 +2,7 @@
 # transform django model into json type
 
 from rest_framework import serializers
-from .models import market_type
+from ..models.stock import market_type
 
 class MarketSerializer(serializers.ModelSerializer):
     class Meta:
