+++ server/db/urls.py
@@ -1,14 +1,14 @@
 from django.urls import path, include
 from rest_framework import routers
-from . import views
+from .views import market
 from . import tests
 
 router = routers.DefaultRouter()
-router.register('market', views.MarketViewSet)
+router.register('market', market.MarketViewSet)
 
 urlpatterns = [
-    path('', views.index, name='index'),
-    path('test/', tests.test, name='test'),
-    path('reset/', tests.reset, name='reset'),
+    path('', market.index, name='index'),
+    path('test/', tests.test, name='test'), # for test
+    path('reset/', tests.reset, name='reset'), # for test
     path('', include(router.urls)),
 ]
