+++ server/server/urls.py
@@ -17,6 +17,6 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path('db/', include('db.urls')),
+    path('', include('db.urls')),
     path('admin/', admin.site.urls),
 ]
